+++ datamanagementapi/src/main/java/org/csi/yucca/storage/datamanagementapi/model/streamOutput/ConfigData.java
@@ -6,7 +6,7 @@ import com.google.gson.annotations.Expose;
 public class ConfigData {
 
     @Expose
-    private Integer idTenant;
+    private Long idTenant;
     @Expose
     private String tenantCode;
     @Expose
@@ -19,7 +19,7 @@ public class ConfigData {
      * @return
      *     The idTenant
      */
-    public Integer getIdTenant() {
+    public Long getIdTenant() {
         return idTenant;
     }
 
@@ -28,7 +28,7 @@ public class ConfigData {
      * @param idTenant
      *     The idTenant
      */
-    public void setIdTenant(Integer idTenant) {
+    public void setIdTenant(Long idTenant) {
         this.idTenant = idTenant;
     }
